Add Supabase integration, ticker functionality, and project setup files
