+++ .~lock.company_tickers_exchange.csv#
@@ -1,0 +1,1 @@
+,DESKTOP-I1I357Q/rachi,DESKTOP-I1I357Q,10.08.2025 01:46,file:///C:/Users/rachi/AppData/Roaming/LibreOffice/4;
