+++ amd_daily.csv
@@ -1,0 +1,101 @@
+date,1. open,2. high,3. low,4. close,5. volume
+2025-08-08,174.0400,176.4800,170.5200,172.7600,68866692
+2025-08-07,166.8350,175.7500,166.7000,172.4000,95448310
+2025-08-06,165.0500,166.1800,157.8000,163.1200,133641835
+2025-08-05,177.5650,177.9900,171.8000,174.3100,88808520
+2025-08-04,174.6050,177.8600,173.5600,176.7800,52951044
+2025-08-01,170.1600,174.4000,166.8200,171.7000,75396130
+2025-07-31,182.0150,182.5000,173.0000,176.3100,71765285
+2025-07-30,175.6050,180.3700,173.8000,179.5100,64820291
+2025-07-29,175.2050,182.3100,174.6800,177.4400,108154838
+2025-07-28,169.0800,174.7000,168.6700,173.6600,68267835
+2025-07-25,163.5050,167.1750,162.3600,166.4700,53432260
+2025-07-24,159.1200,163.9300,158.3600,162.1200,48440112
+2025-07-23,156.3150,159.4485,156.0000,158.6500,41510896
+2025-07-22,156.2000,156.2291,149.3400,154.7200,49028017
+2025-07-21,157.6200,160.3400,156.9200,157.0000,39021130
+2025-07-18,159.5900,160.8300,155.8100,156.9900,48859835
+2025-07-17,161.8100,161.9600,158.6800,160.4100,50605117
+2025-07-16,155.3100,160.3600,152.8500,160.0800,59492760
+2025-07-15,153.7500,158.6800,153.5600,155.6100,93370081
+2025-07-14,145.0900,147.5800,141.9000,146.2400,44718554
+2025-07-11,142.6000,147.4000,141.6000,146.4200,50049179
+2025-07-10,143.0000,145.8200,141.8500,144.1600,61101465
+2025-07-09,138.6900,140.6500,137.5900,138.4100,37013130
+2025-07-08,137.3200,139.1500,135.9100,137.8200,36124612
+2025-07-07,136.5600,137.4400,133.5000,134.8000,37395010
+2025-07-03,139.1100,139.5000,137.3200,137.9100,28645993
+2025-07-02,135.5300,139.7800,135.2200,138.5200,39227986
+2025-07-01,138.7700,140.2300,135.1202,136.1100,55257237
+2025-06-30,144.0000,146.0000,141.0100,141.9000,42972468
+2025-06-27,144.1800,147.7504,141.6300,143.8100,61937853
+2025-06-26,145.9700,146.0000,141.9100,143.6800,58180081
+2025-06-25,140.9400,144.1800,139.4600,143.4000,74607219
+2025-06-24,133.4300,138.7900,132.9300,138.4300,78571997
+2025-06-23,130.4700,133.2500,126.8200,129.5800,65152874
+2025-06-20,129.0000,132.8050,127.5500,128.2400,79930980
+2025-06-18,128.0100,129.2100,125.7700,126.7900,49627112
+2025-06-17,127.9300,130.7000,126.5800,127.1000,86623910
+2025-06-16,118.6350,128.1393,117.7800,126.3900,100968478
+2025-06-13,116.0450,117.8830,115.0600,116.1600,39702742
+2025-06-12,120.6150,122.2500,118.0300,118.5000,44718365
+2025-06-11,124.4400,124.4700,119.8800,121.1400,32261863
+2025-06-10,121.1050,124.0900,120.8000,123.2400,41691908
+2025-06-09,119.1800,122.3699,119.0400,121.7300,55437652
+2025-06-06,117.6850,118.3800,115.8800,116.1900,27042083
+2025-06-05,118.9500,119.2448,114.7100,115.6900,34182902
+2025-06-04,117.4100,119.2950,115.7300,118.5800,31024986
+2025-06-03,114.5000,117.6800,113.2800,117.3100,37161484
+2025-06-02,111.0600,114.8600,111.0100,114.6300,34078358
+2025-05-30,111.5840,112.1100,108.6154,110.7300,36088472
+2025-05-29,115.4850,115.5400,112.1300,113.0300,30572131
+2025-05-28,114.9150,114.9480,112.6600,112.8600,27483741
+2025-05-27,113.1750,115.3950,112.2700,114.5600,37885019
+2025-05-23,108.0850,111.0150,107.6671,110.3100,30867274
+2025-05-22,111.9000,112.6400,110.6100,110.7100,29251957
+2025-05-21,113.8500,116.1000,111.0900,112.0600,43594606
+2025-05-20,114.3500,114.5100,112.0800,113.5100,29030815
+2025-05-19,115.0150,115.5300,114.0600,114.7400,37737373
+2025-05-16,118.0700,118.2400,116.0100,117.1700,43419695
+2025-05-15,117.0200,117.5000,114.0100,114.9900,43843235
+2025-05-14,119.8300,122.5200,116.5900,117.7200,86743700
+2025-05-13,109.6000,113.1277,109.0000,112.4600,55659513
+2025-05-12,110.4150,110.6670,106.9800,108.1200,50543396
+2025-05-09,102.5850,103.5400,101.6000,102.8400,31704261
+2025-05-08,101.9200,105.4400,101.5600,101.7000,54231362
+2025-05-07,100.7500,103.4200,96.8800,100.3600,86648749
+2025-05-06,98.4100,100.0700,97.4400,98.6200,54264254
+2025-05-05,99.0000,102.0500,98.7600,100.5900,41268886
+2025-05-02,98.4000,99.8800,98.3000,98.8000,29957888
+2025-05-01,99.2650,100.1000,96.4500,96.6500,29591507
+2025-04-30,92.8100,97.5500,91.8700,97.3500,29730729
+2025-04-29,95.7550,97.1767,95.3300,96.0600,20700637
+2025-04-28,95.5750,97.4865,94.2700,96.3900,30456683
+2025-04-25,93.2750,97.1000,92.8500,96.6450,28575343
+2025-04-24,91.5900,94.7500,90.3700,94.4700,33681919
+2025-04-23,91.8750,93.7400,90.1200,90.3900,41102589
+2025-04-22,86.1200,87.1700,85.1700,86.2600,35236735
+2025-04-21,86.0150,86.1400,83.7500,85.5600,33855392
+2025-04-17,88.8400,89.0300,86.5150,87.5000,34801437
+2025-04-16,87.6100,90.4700,85.2950,88.2900,62109016
+2025-04-15,94.6600,96.8500,94.6200,95.2900,28116536
+2025-04-14,97.1100,97.3699,93.4500,94.5000,36002970
+2025-04-11,90.5100,94.7600,90.5000,93.4000,57892600
+2025-04-10,91.0000,92.9500,85.0600,88.7000,65210565
+2025-04-09,79.2200,97.9100,78.8700,96.8400,97827352
+2025-04-08,86.1450,88.9300,76.4800,78.2100,69910233
+2025-04-07,80.6800,89.0000,77.7500,83.6400,70515429
+2025-04-04,90.1000,91.0800,83.6700,85.7600,65847021
+2025-04-03,97.7900,99.4499,93.6400,93.8000,50472542
+2025-04-02,101.5200,104.1850,100.7600,102.9600,20793989
+2025-04-01,102.1600,102.8300,100.9300,102.7800,23523230
+2025-03-31,100.2350,102.8200,98.8000,102.7400,30865342
+2025-03-28,105.5250,106.5100,102.4800,103.2200,31069114
+2025-03-27,106.1100,107.6200,105.5499,106.6500,32773714
+2025-03-26,114.1100,115.8999,108.6800,110.1900,39976485
+2025-03-25,113.2650,114.9500,112.7100,114.8100,29077507
+2025-03-24,109.7300,114.8500,109.4000,113.8500,53342002
+2025-03-21,105.3300,107.2500,104.3300,106.4400,30767938
+2025-03-20,105.1200,108.2600,105.1200,107.1400,26871839
+2025-03-19,104.4200,107.9000,101.9600,106.2300,32287496
+2025-03-18,103.8500,104.5500,102.9700,103.5100,27152827
