+++ requirements.txt
@@ -1,3 +1,4 @@
 requests
 pandas
 python-dotenv
+supabase
